Renamed the casing of file names to slug case following internal discussion on naming convention.
