+++ template/src/actions/index.js
@@ -1,4 +1,4 @@
-import {{pascalcase name}}Actions from './Actions';
+import {{pascalcase name}}Actions from './actions';
 
 export default {{pascalcase name}}Actions;
 export { {{pascalcase name}}Actions };
