+++ template/src/components/toggleButton/index.js
@@ -1,4 +1,4 @@
-import ToggleButton from './ToggleButton';
+import ToggleButton from './toggle-button';
 
 export default ToggleButton;
 export { ToggleButton };
