+++ template/src/plugin.js
