+++ template/src/stores/index.js
@@ -1,4 +1,4 @@
-import {{pascalcase name}}Store from './Store';
+import {{pascalcase name}}Store from './store';
 
 export default {{pascalcase name}}Store;
 export { {{pascalcase name}}Store };
