+++ template/src/stores/store.spec.js
